Add support for Debian Wheezy and HamServerPi
This commit adds support for Debian Wheezy and its derivates. A lot of servers
on the HAMNET use HamServerPi, which sadly still depends on Debian Wheezy.
Wheezy comes with Python 3.2 and SysVInit preinstalled. This commit is a first
approach to adding support to those legacy program versions.

Good luck with your bachelors thesis!

Signed-off-by: Anton Würfel DL8AW <info@dl8aw.de>